AIフォルダ作成、微調整
