+++ Assets/YagiWorks/AI.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: fbfc73fc7ac04774daea9a2910491cc8
+folderAsset: yes
+timeCreated: 1448790657
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
